# backend/docs/swagger.yaml
definitions:
...
  common.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      timestamp:
        example: "2025-08-29T10:00:00Z"
        type: string
    type: object
  common.PaginationResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      message:
        example: success
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      timestamp:
        example: "2025-08-29T10:00:00Z"
        type: string
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  common.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: success
        type: string
      timestamp:
        example: "2025-08-29T10:00:00Z"
        type: string
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.CreateProjectRequest:
    properties:
      requirements:
        example: 项目需求描述
        type: string
    required:
    - requirements
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.ProjectInfo:
    properties:
      backend_port:
        example: 8080
        type: integer
      created_at:
        type: string
      description:
        example: 项目描述
        type: string
      frontend_port:
        example: 3000
        type: integer
      guid:
        example: e080335a93d0456ba9b65ab407710e55
        type: string
      name:
        example: 项目名称
        type: string
      preview_url:
        example: http://guid.app-maker.localhost
        type: string
      project_path:
        example: /path/to/project
        type: string
      requirements:
        example: 项目需求
        type: string
      status:
        example: in_progress
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UserInfo'
      user_id:
        example: USER_00000000001
        type: string
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.UpdateProjectRequest:
    properties:
      ai_model:
        example: glm-4.6
        type: string
      cli_tool:
        enum:
        - claude-code
        - qwen-code
        - gemini
        example: claude-code
        type: string
      description:
        example: 项目描述
        type: string
      model_api_url:
        example: https://open.bigmodel.cn/api/anthropic
        type: string
      model_provider:
        enum:
        - ollama
        - zhipu
        - anthropic
        - openai
        - vllm
        example: zhipu
        type: string
      name:
        example: 项目名称
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
  github_com_lighthought_app-maker_backend_internal_models.UserInfo:
    properties:
      created_at:
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: varchar(50)
        type: string
      role:
        example: user
        type: string
      status:
        example: active
        type: string
      username:
        example: username
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/preview/{token}:
    get:
      consumes:
      - application/json
      description: 通过分享令牌访问项目预览（无需认证）
      parameters:
      - description: 预览令牌
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到预览页面
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: 令牌不存在或已过期
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 通过令牌访问预览
      tags:
      - 项目管理
  /api/v1/projects:
    get:
      consumes:
      - application/json
      description: 获取项目列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 项目状态
        enum:
        - draft
        - in_progress
        - completed
        - failed
        in: query
        name: status
        type: string
      - description: 搜索关键词
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取项目列表成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.PaginationResponse'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.ProjectInfo'
                        type: array
                    type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取项目列表
      tags:
      - 项目管理
    post:
      consumes:
      - application/json
      description: 创建新项目
      parameters:
      - description: 项目创建请求
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 项目创建成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.ProjectInfo'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 创建项目
      tags:
      - 项目管理
  /api/v1/projects/{guid}:
    delete:
      consumes:
      - application/json
      description: 删除指定项目
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 项目删除成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除项目
      tags:
      - 项目管理
    get:
      consumes:
      - application/json
      description: 根据项目ID获取项目详细信息
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取项目成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.ProjectInfo'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取项目信息
      tags:
      - 项目管理
    put:
      consumes:
      - application/json
      description: 更新指定项目的配置信息
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      - description: 项目更新请求
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_lighthought_app-maker_backend_internal_models.ProjectInfo'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新项目
      tags:
      - 项目管理
  /api/v1/projects/{guid}/deploy:
    post:
      consumes:
      - application/json
      description: 部署指定项目
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 项目部署成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 部署项目
      tags:
      - 项目管理
...
  /api/v1/projects/{guid}/preview-link:
    post:
      consumes:
      - application/json
      description: 为项目生成可分享的预览链接
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      - default: 7
        description: 过期天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功生成分享链接
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 生成预览分享链接
      tags:
      - 项目管理
  /api/v1/projects/{guid}/stages:
    get:
      consumes:
      - application/json
      description: 获取指定项目的开发阶段信息
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取项目开发阶段
      tags:
      - 开发阶段
  /api/v1/projects/download/{guid}:
    get:
      consumes:
      - application/json
      description: 将项目文件打包为zip并下载
      parameters:
      - description: 项目GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: 项目文件zip包
          schema:
            type: file
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - Bearer: []
      summary: 下载项目文件
      tags:
      - 项目管理
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"